requirements.txt
beautifulsoup4==4.10.0
blis==0.7.4
bs4==0.0.1
catalogue==2.0.6
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.6
click==8.0.2
cymem==2.0.5
Deprecated==1.2.13
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.1.0/en_core_web_sm-3.1.0-py3-none-any.whl
idna==3.2
Jinja2==3.0.2
joblib==1.1.0
lxml==4.6.3
MarkupSafe==2.0.1
murmurhash==1.0.5
nltk==3.6.3
numpy==1.21.2
packaging==21.0
pandas==1.3.3
pathy==0.6.0
preshed==3.0.5
psutil==5.8.0
pycparser==2.20
pydantic==1.8.2
PyGithub==1.55
PyJWT==2.2.0
PyNaCl==1.4.0
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.3
regex==2021.10.8
requests==2.26.0
six==1.16.0
smart-open==5.2.1
soupsieve==2.2.1
spacy==3.1.3
spacy-legacy==3.0.8
srsly==2.4.1
stanfordcorenlp==3.9.1.1
thinc==8.0.10
tqdm==4.62.3
typer==0.4.0
typing-extensions==3.10.0.2
urllib3==1.26.7
wasabi==0.8.2
wrapt==1.13.2
